Position is a CGpoint, which are CGFloats

+++ swiftgame/classes/parallax/ParallaxSprite.swift
@@ -48,7 +48,7 @@ class ParallaxSprite: SKSpriteNode {
     
     func update() {
         if(self.parallaxSpeed > 0){
-            self.position.x -= Float(self.parallaxSpeed);
+            self.position.x -= CGFloat(self.parallaxSpeed);
             
             if(self.position.x < -self.image1.size.width) {
                 self.position.x = 0;
